+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/JsonRpcError.java
@@ -39,8 +39,7 @@ public enum JsonRpcError {
       "The method eth_sendTransaction is not supported. Use eth_sendRawTransaction to send a signed transaction to Besu."),
   ETH_SEND_TX_ALREADY_KNOWN(-32000, "Known transaction"),
   ETH_SEND_TX_REPLACEMENT_UNDERPRICED(-32000, "Replacement transaction underpriced"),
-  ETH_FORBIDDEN_UNPROTECTED_TRANSACTION(-32000, "Forbidden unprotected transactions"),
-  // P2P related errors
+  // P2P related errors,
   P2P_DISABLED(-32000, "P2P has been disabled. This functionality is not available"),
   P2P_NETWORK_NOT_RUNNING(-32000, "P2P network is not running"),
 
@@ -65,6 +64,8 @@ public enum JsonRpcError {
   TX_FEECAP_EXCEEDED(-32000, "Transaction fee cap exceeded"),
   REVERT_ERROR(-32000, "Execution reverted"),
   GAS_PRICE_MUST_BE_ZERO(-3200, "gasPrice must be set to zero on a GoQuorum compatible network"),
+  TRANSACTION_WITHOUT_REPLAY_PROTECTION(
+      -32000, "Transaction missing replay protection where it is required"),
 
   // Miner failures
   COINBASE_NOT_SET(-32010, "Coinbase not set. Unable to start mining without a coinbase"),